% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{apriori.density}
\alias{apriori.density}
\title{Returns a standard a-priori density function}
\usage{
apriori.density(density.id)
}
\arguments{
\item{density.id}{'U', 'T' or 'HT', depending on which a-priori density should
be returned}
}
\value{
Uniform (U), triangle (T) or half triangle (HT) a-priori density function
}
\description{
Let U denote the uniform distribution on \eqn{[0,1]}, T the triangle distribution
with center point \eqn{0.5} and full support on \eqn{[0,1]} and HT the half-triangle distribution
with center point \eqn{0.5} and support on \eqn{[0.25, 0.75]}. This function returns one of
these a-priori densities in dependence of its input parameter. These three
a-priori distributions are deemed as helpful in experimenting with
the Schuirmann constant.
}
\examples{
 
rho <- apriori.density('T')
r <- seq(0, 1, length = 100)
plot(r, rho(r), type = 'l')
}
