% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_point.R
\name{n.point}
\alias{n.point}
\title{Exact sample size; point mean-difference}
\usage{
n.point(alpha, theta1, theta2, theta0, sigma, pwr)
}
\arguments{
\item{alpha}{type I error}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{theta0}{mean-difference}

\item{sigma}{standard deviation of the measurements}

\item{pwr}{Exact power of the TOST}
}
\value{
sample size in each group
}
\description{
Exact sample size calculation of a balanced, unpaired TOST. The result n are
the samples in one group. The true mean-difference is assumed to be a point,
i.e. fixed value.
}
\examples{
pwr <- power.point(alpha = 0.05, theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, n = 15)
n.point(alpha = 0.05, theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, pwr = pwr)
}
