% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{n.weighted}
\alias{n.weighted}
\title{Exact sample size; weighted mean-difference}
\usage{
n.weighted(alpha, theta1, theta2, density, sigma, pwr)
}
\arguments{
\item{alpha}{type I error}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{density}{a-priori density for the mean-difference}

\item{sigma}{standard deviation of the measurements}

\item{pwr}{Exact power of the TOST}
}
\value{
sample size in each group
}
\description{
Exact sample size calculation of a balanced, unpaired TOST, where the true
mean-difference follows a prescribed a-priori distribution. The lowest sample
size n that exceeds the target power pwd is calculated using a standard binary
search algorithm. The result n is the sample size in one group.
}
\examples{
rho <- apriori.density('U')
pwr <- power.weighted(alpha = 0.05, theta1 = -2, theta2 = 2, density = rho, sigma = 1, n = 22)
n.weighted(alpha = 0.05, theta1 = -2, theta2 = 2, density = rho, sigma = 1, pwr = pwr)
}
