% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_point.R
\name{nabla.point}
\alias{nabla.point}
\title{nabla; point mean-difference}
\usage{
nabla.point(alpha, pwr, r)
}
\arguments{
\item{alpha}{type I error}

\item{pwr}{asymptotic power of the TOST}

\item{r}{relative mean-difference}
}
\value{
\eqn{\nabla}{nabla} for a fixed point mean-difference
}
\description{
The equality
\deqn{{power}_{asymp}(\nabla, r, \alpha) = pwr}{
power_asymp(nabla, r, alpha) = pwr}
is solved for \eqn{\nabla}{nabla} in dependence of \eqn{\alpha}{alpha}, \eqn{pwr}
and \eqn{r}.
}
\examples{
nabla <- nabla.point(alpha = 0.05, pwr = 0.8, r = 0.55)
power.point.asymp(alpha = 0.05, r = 0.55, nabla = nabla)
}
