% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformation.R
\name{nabla.to.standard}
\alias{nabla.to.standard}
\title{TOST sensitivity to standard parameterization}
\usage{
nabla.to.standard(r, nabla, theta1, theta2, n)
}
\arguments{
\item{r}{relative mean-difference}

\item{nabla}{TOST sensitivity index}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{n}{sample size in each group}
}
\value{
A list with the transformed parameters in standard parameterization, i.e.
\eqn{(\theta_1, \theta_2, \theta_0, \sigma, n)}{(theta1, theta2, theta0, sigma, n)}
}
\description{
The TOST sensitivity parameterization is transformed to the standard parameterization.
}
\examples{
l <- nabla.to.standard(r = 0.75, nabla = 10, theta1 = 0, theta2 = 4, n = 20)
standard.to.nabla(l$theta1,l$theta2,l$theta0, l$sigma, l$n)
}
