% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{nabla.weighted}
\alias{nabla.weighted}
\title{nabla; weighted mean-difference}
\usage{
nabla.weighted(alpha, pwr, density)
}
\arguments{
\item{alpha}{type I error}

\item{pwr}{asymptotic weighted power of the TOST}

\item{density}{a-priori density for the mean-difference}
}
\value{
\eqn{\nabla}{nabla} for a weighted mean-difference
}
\description{
The equality
\deqn{power_{asymp, density}(\nabla, \alpha) = pwr}{
power_asymp_density(nabla, alpha) = pwr}
is solved for \eqn{\nabla}{nabla} in dependence of \eqn{\alpha}{alpha},
\eqn{pwr}{pwr} and \eqn{density}{density}.
}
\examples{
rho <- apriori.density('U')
nabla <- nabla.weighted(alpha = 0.05, pwr = 0.8, density = rho)
power.weighted.asymp(alpha = 0.05, density = rho, nabla = nabla)
}
