% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_point.R
\name{power.point.asymp}
\alias{power.point.asymp}
\title{Asymptotical TOST power; point mean-difference}
\usage{
power.point.asymp(alpha, r, nabla)
}
\arguments{
\item{alpha}{type I error}

\item{r}{relative mean-difference}

\item{nabla}{TOST sensitivity index}
}
\value{
Asymptotic power of the TOST for a fixed point mean-difference
}
\description{
Calculates the asymptotical power of an unpaired, balanced TOST for a fixed
point mean-difference.
}
\examples{
power.point.nabla(alpha = 0.05, r = 2/3, nabla = 10, n = 10)
power.point.nabla(alpha = 0.05, r = 2/3, nabla = 10, n = 50)
power.point.asymp(alpha = 0.05, r = 2/3, nabla = 10)
}
