% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_point.R
\name{power.point.nabla}
\alias{power.point.nabla}
\title{Exact power; point mean-difference; sensitivity parameterization}
\usage{
power.point.nabla(alpha, r, nabla, n)
}
\arguments{
\item{alpha}{type I error}

\item{r}{relative mean-difference}

\item{nabla}{TOST sensitivity index}

\item{n}{sample size in each group}
}
\value{
Exact power of the TOST for a fixed point mean-difference
}
\description{
Calculates the exact power of an unpaired, balanced TOST with n samples in
each group in the TOST sensitivity parameterization. The true mean-difference
is assumed to be a point, i.e. fixed value.
}
\examples{
power.point(alpha = 0.05, theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, n = 15)
l <- standard.to.nabla(theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, n = 15)
power.point.nabla(alpha = 0.05, r = l$r, nabla = l$nabla, n = 15)
}
