% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_point.R
\name{power.point}
\alias{power.point}
\title{Exact power; point mean-difference; standard parameterization}
\usage{
power.point(alpha, theta1, theta2, theta0, sigma, n)
}
\arguments{
\item{alpha}{type I error}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{theta0}{mean-difference}

\item{sigma}{standard deviation of the measurements}

\item{n}{sample size in each group}
}
\value{
Exact power of the TOST for a fixed point mean-difference
}
\description{
Calculates the exact power of an unpaired, balanced TOST with n samples in
each group. The true mean-difference is assumed to be a point, i.e. fixed value.
}
\examples{
pwr <- power.point(alpha = 0.05, theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, n = 15)
n.point(alpha = 0.05, theta1 = -1, theta2 = 2, theta0 = 1, sigma = 1, pwr = pwr)
}
