% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{power.weighted.asymp}
\alias{power.weighted.asymp}
\title{Asymptotical weighted power}
\usage{
power.weighted.asymp(alpha, density, nabla)
}
\arguments{
\item{alpha}{type I error}

\item{density}{a-priori density for the mean-difference}

\item{nabla}{TOST sensitivity index}
}
\value{
Asymptotic power of the TOST for a weighted mean-difference
}
\description{
Calculates the asymptotical power of an unpaired, balanced TOST
for a weighted a-priori mean-difference.
}
\examples{
rho <- apriori.density('HT')
power.weighted.nabla(alpha = 0.05, density = rho, nabla = 7, n = 10)
power.weighted.nabla(alpha = 0.05, density = rho, nabla = 7, n = 50)
power.weighted.asymp(alpha = 0.05, density = rho, nabla = 7)
}
