% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{power.weighted.nabla}
\alias{power.weighted.nabla}
\title{Exact weighted power; sensitivity parameterization}
\usage{
power.weighted.nabla(alpha, density, nabla, n)
}
\arguments{
\item{alpha}{type I error}

\item{density}{a-priori density for the mean-difference}

\item{nabla}{TOST sensitivity index}

\item{n}{sample size in each group}
}
\value{
Exact power of the TOST for a weighted mean-difference
}
\description{
Calculates the exact power of an unpaired, balanced TOST with
n samples in each group, where the true mean-difference is modeled with an
a-priori density rather than a fixed mean-difference. The calculation is performed
in the TOST sensitivity parameterization.
}
\examples{
rho <- apriori.density('U')
power.weighted.nabla(alpha = 0.05, density = rho, nabla = 8, n = 22)
## Since we are performing a weighted power calculation, the r respectively 
## theta0 value is not needed. Thus, the following calculation does not depend
## on r.
l <- nabla.to.standard(r = 0.99, nabla = 8, theta1 = -1, theta2 = 4, n = 22)
power.weighted(alpha = 0.05, theta1 = -1, theta2 = 4, density = rho, sigma = l$sigma, n = 22)
}
