% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_diff_weighted.R
\name{power.weighted}
\alias{power.weighted}
\title{Exact weighted power; standard parameterization}
\usage{
power.weighted(alpha, theta1, theta2, density, sigma, n)
}
\arguments{
\item{alpha}{type I error}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{density}{a-priori density for the mean-difference}

\item{sigma}{standard deviation of the measurements}

\item{n}{sample size in each group}
}
\value{
Exact power of the TOST for a weighted true mean-difference
}
\description{
Calculates the exact power of an unpaired, balanced TOST with
n samples in each group, where the true mean-difference is modeled with an
a-priori density rather than a fixed mean-difference.
}
\examples{
rho <- apriori.density('U')
pwr <- power.weighted(alpha = 0.05, theta1 = -2, theta2 = 2, density = rho, sigma = 1, n = 22)
n.weighted(alpha = 0.05, theta1 = -2, theta2 = 2, density = rho, sigma = 1, pwr = pwr)
}
