% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schuirmann_const.R
\name{schuirmann.constant}
\alias{schuirmann.constant}
\title{Calculation of the schuirmann-constant}
\usage{
schuirmann.constant(alpha, pwr, density)
}
\arguments{
\item{alpha}{type I error}

\item{pwr}{asymptotic weighted power of the TOST}

\item{density}{a-priori density for the mean-difference}
}
\value{
The Schuirmann constant
}
\description{
Calculation of the schuirmann-constant, cf. \insertCite{palmes2020}{schuirmann.constant}
}
\examples{
rho <- apriori.density('U')
theta1 <- 0
theta2 <- 3
s <- schuirmann.constant(alpha = 0.05, pwr = 0.8, density = rho)
theta0 <- (1-s)*theta1 + s*theta2 
n <- n.point(alpha = 0.05, theta1 = theta1, theta2 = theta2, theta0 = theta0, sigma = 1, pwr = 0.8)
## Note that the weighted power is calculated with a sample size that was found for a
## point-difference. Due to the newly introduced duality concept it is -nevertheless-
## very close to the target power 0.8. 
power.weighted(alpha = 0.05, theta1 = theta1, theta2 = theta2, density = rho, sigma = 1, n = n)
}
\references{
\insertAllCited{}
}
