% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schuirmann_const.R
\name{schuirmann.constant.uniform}
\alias{schuirmann.constant.uniform}
\title{Calculation of the uniform schuirmann-constant}
\usage{
schuirmann.constant.uniform(alpha, pwr, theta1 = 0, theta2 = 1)
}
\arguments{
\item{alpha}{type I error}

\item{pwr}{asymptotic weighted power of the TOST}

\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}
}
\value{
The Schuirmann constant of the uniform a-priori distribtion is returned.
If \eqn{[\theta_1,  \theta_2]}{[theta1,  theta2]} differs from \eqn{[0,1]}, then
the appropriately scaled true-mean difference
\deqn{\theta_0 = (S-1) \cdot \theta_1 + S \cdot \theta_2}{
theta0 = (S-1)*theta1 + S*theta2}
is returned.
}
\description{
Calculation of the uniform schuirmann-constant, cf. \insertCite{palmes2020}{schuirmann.constant}
}
\details{
The uniform density is assumed as a-priori distribution. This density is
unique with the property that each possible mean-difference is equally weighted.
It represents the lack of any information about the true mean-difference. Due
to its importance, this special setting is included as a separate function for
convenience. Technical, this function is merely a wrapper that calls the
schuirmann.constant function with the uniform distribution as a-priori density.
}
\examples{
rho <- apriori.density('U')
theta1 <- 0
theta2 <- 3
s <- schuirmann.constant(alpha = 0.05, pwr = 0.8, density = rho)
(1-s)*theta1 + s*theta2
schuirmann.constant.uniform(alpha = 0.05, pwr = 0.8, theta1 = theta1, theta2 = theta2) 
}
\references{
\insertAllCited{}
}
