% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformation.R
\name{standard.to.nabla}
\alias{standard.to.nabla}
\title{standard to TOST sensitivity parameterization}
\usage{
standard.to.nabla(theta1, theta2, theta0, sigma, n)
}
\arguments{
\item{theta1}{lower limit of equivalence interval}

\item{theta2}{upper limit of equivalence interval}

\item{theta0}{mean-difference}

\item{sigma}{standard deviation of the measurements}

\item{n}{sample size in each group}
}
\value{
A list with the transformed parameters in TOST sensitivity parameterization,
i.e. \eqn{(r, \nabla, \theta_1, \theta_2, n)}{(r, nabla, theta1, theta2, n)}
}
\description{
The standard parameterization is transformed to the TOST sensitivity parameterization.
}
\examples{
l <- nabla.to.standard(r = 0.75, nabla = 10, theta1 = 0, theta2 = 4, n = 20)
standard.to.nabla(l$theta1,l$theta2,l$theta0, l$sigma, l$n)
}
